Refactor `tf_proxy` to use `grpc_attestation`
