+++ experimental/tf_proxy/server/Cargo.toml
@@ -10,6 +10,7 @@ async-stream = "*"
 anyhow = "*"
 env_logger = "*"
 futures = "*"
+grpc_attestation = { path = "../../../grpc_attestation/" }
 log = "*"
 oak_functions_abi = { path = "../../../oak_functions/abi" }
 oak_remote_attestation = { path = "../../../remote_attestation/rust/" }
