+++ experimental/tf_proxy/server/build.rs
@@ -14,22 +14,9 @@
 // limitations under the License.
 //
 
-use oak_utils::{generate_grpc_code, CodegenOptions, ExternPath};
+use oak_utils::{generate_grpc_code, CodegenOptions};
 
 fn main() -> Result<(), Box<dyn std::error::Error>> {
-    generate_grpc_code(
-        "../../../",
-        &["oak_functions/proto/server.proto"],
-        CodegenOptions {
-            build_client: false,
-            build_server: true,
-            extern_paths: vec![ExternPath::new(
-                ".oak.remote_attestation",
-                "::oak_remote_attestation::proto",
-            )],
-        },
-    )?;
-
     generate_grpc_code(
         // We would normally use `../../..` as the `proto_path`, but that would require modifying
         // the import statements of the third party protos.
